Fix: Topology instance detection
